rename _getPackage -> getPackages to follow naming convention
