+++ scripts/getPackages.js
