+++ scripts/watch.js
@@ -13,7 +13,7 @@ const fs = require('fs');
 const {execSync} = require('child_process');
 const path = require('path');
 const chalk = require('chalk');
-const getPackages = require('./_getPackages');
+const getPackages = require('./getPackages');
 
 const BUILD_CMD = `node ${path.resolve(__dirname, './build.js')}`;
